renamed to SpeechNavigation + InitialConsonants
