+++ texas-hearing-institute/App.tsx
@@ -1,5 +1,5 @@
 
-import SpeechSetup from "./pages/Home";
+import SpeechSetup from "./pages/SpeechNavigation";
 function App() {
   return <SpeechSetup/>
 }
